requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
streamlit
langchain-chroma
langchain-google-genai
pypdf2
uuid
google-generativeai